+++ ts/packages/cli/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @composio/cli
 
+## 0.2.4-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
 ## 0.2.4
 
 ### Patch Changes