+++ ts/packages/providers/google/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @composio/google
 
+## 0.6.5-alpha.1
+
+### Patch Changes
+
+- Updated dependencies
 ## 0.6.5
 
 ### Patch Changes